Give Hind SHP Shadow
There are 2 issues with this.

I forgot the fact that voxel had a different scale, and i didn't wanna
redo the whole thing so i scaled the shadow too.

It looks wierd that the shadow of units it is carrying is rendered with
voxel renderer, so it is rendered at a different position than SHP one.