Change Structure real height's to 0.
This means you need to click base of a building to interact with them. This may end up worse when it comes to selling wrong things, but it makes it so selection boxes cannot overlap. Meaning and wrong selling is player's fault and hopefully player's will get used to it.